+++ src/main/kotlin/org/xbery/artbeams/products/domain/EditedProduct.kt
@@ -7,7 +7,10 @@ data class EditedProduct(
     val id: String,
     val slug: String,
     val title: String,
+    val subtitle: String?,
     val fileName: String?,
+    val listingImage: String?,
+    val image: String?,
     val confirmationMailingGroupId: String?,
     val mailingGroupId: String?
 )