+++ src/main/kotlin/org/xbery/artbeams/products/domain/Product.kt
@@ -11,8 +11,13 @@ data class Product(
     override val common: AssetAttributes,
     val slug: String,
     val title: String,
+    val subtitle: String?,
     /** File name within {@link FileData} in media module. Filled if this is digital product that can be downloaded. */
     val fileName: String?,
+    /** Image for listing view. */
+    val listingImage: String?,
+    /** Image for detail view. */
+    val image: String?,
     /** Intermediate mailing group id for subscription confirmation related to this product. */
     val confirmationMailingGroupId: String?,
     /** Mailing group id for subscription related to this product. */
@@ -23,19 +28,22 @@ data class Product(
             common = this.common.updatedWith(userId),
             slug = edited.slug,
             title = edited.title,
+            subtitle = edited.subtitle,
             fileName = edited.fileName,
+            listingImage = edited.listingImage,
+            image = edited.image,
             confirmationMailingGroupId = edited.confirmationMailingGroupId,
             mailingGroupId = edited.mailingGroupId
         )
     }
 
     fun toEdited(): EditedProduct {
-        return EditedProduct(this.id, this.slug, this.title, this.fileName, this.confirmationMailingGroupId, this.mailingGroupId)
+        return EditedProduct(this.id, this.slug, this.title, this.subtitle, this.fileName, this.listingImage, this.image, this.confirmationMailingGroupId, this.mailingGroupId)
     }
 
     companion object {
         val Empty = Product(
-            AssetAttributes.Empty, "", "New product", null, null, null
+            AssetAttributes.Empty, "", "New product", null, null, null, null, null, null
         )
     }
 }