+++ src/main/kotlin/org/xbery/artbeams/userproducts/domain/UserProductInfo.kt
@@ -8,6 +8,7 @@ package org.xbery.artbeams.userproducts.domain
 data class UserProductInfo(
     val id: String,
     val title: String,
+    val subtitle: String?,
     val slug: String,
-    val subtitle: String?
+    val listingImage: String?
 )