+++ src/main/resources/sql/create_tables.sql
@@ -86,7 +86,10 @@ CREATE TABLE products (
 	modified_by VARCHAR(40) DEFAULT NULL,
 	slug VARCHAR(128) DEFAULT NULL,
 	title VARCHAR(128) DEFAULT NULL,
+	subtitle VARCHAR(256) DEFAULT NULL,
 	filename VARCHAR(128) DEFAULT NULL,
+	listing_image VARCHAR(128) DEFAULT NULL,
+	image VARCHAR(128) DEFAULT NULL,
 	confirmation_mailing_group_id VARCHAR(128) DEFAULT NULL,
 	mailing_group_id VARCHAR(128) DEFAULT NULL
 );